requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
#dj-static==0.0.6
# Install directly from github repo:
git+https://github.com/kennethreitz/dj-static.git
# Django color field
git+https://github.com/jaredly/django-colorfield.git
django-auth-ldap==1.2.0
django-compressor==1.4
django-crispy-forms==1.4.0
django-easy-pdf
django-image-cropping==0.8.2
django-recurrence==1.0.1
django-suit==0.2.9
django-tinymce==1.5.2
easy-thumbnails==2.0.1
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
python-ldap==2.4.15
pytz==2014.2
reportlab==2.7 # old version for python 2.6 compatibility in RHEL6
requests==2.3.0
six==1.6.1
static==1.0.2
static3==0.5
uWSGI==2.0.4
wsgiref==0.1.2
git+https://github.com/chrisglass/xhtml2pdf.git 
